NP-435 moved formcard to a directory
